Remove user_id key as required on front end from all routes

+++ user/user-routes.js
@@ -61,13 +61,15 @@ router.put("/:id", (req, res) => {
 
 // POST to user's values
 router.post("/:id/values", (req, res) => {
-  const { value_id, user_id } = req.body;
+  const { value_id } = req.body;
+  const { id } = req.params;
 
-  if (!value_id || !user_id) {
+  if (!value_id) {
     res
       .status(400)
-      .json({ message: "You must submit a value id and user id." });
+      .json({ message: "You must submit a value id." });
   } else {
+    req.body.user_id = id;
     User.addValue(req.body)
       .then(id => {
         res.status(201).json(id);
@@ -82,13 +84,15 @@ router.post("/:id/values", (req, res) => {
 
 // POST to user's projects
 router.post("/:id/projects", (req, res) => {
-  const { project_name, user_id } = req.body;
+  const { project_name } = req.body;
+  const { id } = req.params;
 
-  if (!project_name || !user_id) {
+  if (!project_name) {
     res
       .status(400)
-      .json({ message: "You must submit a project name and user id." });
+      .json({ message: "You must submit a project name." });
   } else {
+    req.body.user_id = id;
     User.addProject(req.body)
       .then(id => {
         res.status(201).json(id);
@@ -102,13 +106,15 @@ router.post("/:id/projects", (req, res) => {
 });
 // POST to user's journal
 router.post("/:id/journal", (req, res) => {
-  const { journal_entry, user_id } = req.body;
+  const { journal_entry } = req.body;
+  const { id } = req.params;
 
-  if (!journal_entry || !user_id) {
+  if (!journal_entry) {
     res
       .status(400)
       .json({ message: "You must submit a journal entry and user id." });
   } else {
+    req.body.user_id = id;
     User.addJournal(req.body)
       .then(journal => {
         res.status(201).json(journal);
@@ -123,17 +129,18 @@ router.post("/:id/journal", (req, res) => {
 
 // PUT a user's value
 router.put("/:id/values", (req, res) => {
-  const { old_value_id, value_id, user_id } = req.body;
+  const { old_value_id, value_id } = req.body;
+  const { id } = req.params;
 
-  if (!old_value_id || !value_id || !user_id) {
+  if (!old_value_id || !value_id) {
     res
       .status(400)
-      .json("You must submit an old value id, new value id, and user id.");
+      .json("You must submit an old value id and a new value id.");
   } else {
-    User.getUserValueByValueId(old_value_id, user_id)
+    User.getUserValueByValueId(old_value_id, id)
       .then(value => {
         if (value) {
-          User.updateValue(old_value_id, user_id, value_id).then(updated => {
+          User.updateValue(old_value_id, id, value_id).then(updated => {
             if (updated > 0) {
               return User.getValueById(value_id).then(newValue => {
                 res.status(200).json(newValue);
@@ -158,41 +165,51 @@ router.put("/:id/values", (req, res) => {
 // PUT a user's project
 router.put("/:id/projects", (req, res) => {
   const { project_name, id } = req.body;
+  let userId = req.params.id;
 
   if (!project_name || !id) {
     res
       .status(400)
       .json({ message: "You must submit a project name and project id." });
   } else {
-    User.updateProject(id, req.body)
-      .then(updated => {
-        if (updated > 0) {
-          return User.getProjectById(id).then(project => {
-            res.status(201).json(project);
+    User.getProjectById(id).then(project => {
+      if (userId === `${project.user_id}`) {
+        User.updateProject(id, req.body)
+        .then(updated => {
+          if (updated > 0) {
+            return User.getProjectById(id).then(project => {
+              res.status(201).json(project);
+            });
+          }
+        })
+        .catch(err => {
+          res.status(500).json({
+            message: "There was an error updating the project."
           });
-        }
-      })
-      .catch(err => {
-        res.status(500).json({
-          message: "There was an error updating the project."
         });
-      });
+      } else {
+        res.status(403).json({ message: "This project id does not belong to this user."})
+      }
+    })
+
   }
 });
 
 // PUT a user's journal
 router.put("/:id/journal", (req, res) => {
-  const { journal_entry, user_id } = req.body;
+  const { journal_entry } = req.body;
+  const { id } = req.params;
 
-  if (!journal_entry || !user_id) {
+  if (!journal_entry) {
     res
       .status(400)
-      .json({ message: "You must submit a journal entry and user id." });
+      .json({ message: "You must submit a journal entry." });
   } else {
-    User.updateJournal(user_id, req.body)
+    req.body.user_id = id;
+    User.updateJournal(id, req.body)
       .then(updated => {
         if (updated > 0) {
-          return User.getUserJournalByUserId(user_id).then(updates => {
+          return User.getUserJournalByUserId(id).then(updates => {
             let [journal] = updates;
             res.status(201).json(journal);
           });
@@ -209,25 +226,33 @@ router.put("/:id/journal", (req, res) => {
 // DELETE a user's project
 router.delete("/:id/projects", (req, res) => {
   let { id } = req.body;
+  let userId = req.params.id;
 
   if (!id) {
     res.status(400).json({ message: "You must submit a project id." });
   } else {
-    User.deleteProject(id)
-      .then(deleted => {
-        if (deleted > 0) {
+    User.getProjectById(id).then(project => {
+      if (userId === `${project.user_id}`) {
+        User.deleteProject(id)
+        .then(deleted => {
+          if (deleted > 0) {
+            res
+              .status(200)
+              .json({ message: "The project has been successfully deleted." });
+          } else {
+            res.status(404).json({ message: "That project id does not exist." });
+          }
+        })
+        .catch(err => {
           res
-            .status(200)
-            .json({ message: "The project has been successfully deleted." });
-        } else {
-          res.status(404).json({ message: "That project id does not exist." });
-        }
-      })
-      .catch(err => {
-        res
-          .status(500)
-          .json({ message: "There was an error deleting the project." });
-      });
+            .status(500)
+            .json({ message: "There was an error deleting the project." });
+        });
+      } else {
+        res.status(403).json({ message: "This project id does not belong to this user."})
+      }
+    })
+
   }
 });
 
@@ -252,14 +277,15 @@ router.delete("/:id/journal", (req, res) => {
 
 // DELETE a user's value
 router.delete("/:id/values", (req, res) => {
-  const { value_id, user_id } = req.body;
+  const { value_id } = req.body;
+  const { id } = req.params;
 
-  if (!value_id || !user_id) {
+  if (!value_id) {
     res
       .status(400)
       .json({ message: "You must submit a value id and user id." });
   } else {
-    User.deleteUserValue(value_id, user_id)
+    User.deleteUserValue(value_id, id)
       .then(deleted => {
         if (deleted > 0) {
           return res
